Split up regex replacements into separate files so we can test them individually.
