+++ src/FakerDotNet/FakerUtils/RegexifyReplacers/IRegexifyReplacer.cs
@@ -1,0 +1,7 @@
+namespace FakerDotNet.FakerUtils.RegexifyReplacers
+{
+    internal interface IRegexifyReplacer
+    {
+        string Run(string pattern);
+    }
+}
